% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_codedoc.R
\docType{package}
\name{codedoc-package}
\alias{codedoc}
\alias{codedoc-package}
\title{codedoc: Generate Code Documentation}
\description{
Allows generation of documentation from specifically
formatted code comments.

\href{https://github.com/FinnishCancerRegistry/codedoc/actions/workflows/R-CMD-check.yaml}{\if{html}{\figure{https://github.com/FinnishCancerRegistry/codedoc/actions/workflows/R-CMD-check.yaml/badge.svg}}}
}
\section{Recommended installation}{
\if{html}{\out{<div class="sourceCode r">}}\preformatted{devtools::install_github(
  "FinnishCancerRegistry/codedoc",
  ref = readline("enter latest tag on github: ")
)
}\if{html}{\out{</div>}}
}

\section{Example}{
\if{html}{\out{<div class="sourceCode r">}}\preformatted{block_df <- codedoc::extract_keyed_comment_blocks(
  text_file_paths = codedoc::example_text_file_path("r_script.R")
)
print(block_df)
}\if{html}{\out{</div>}}
}

\section{News}{

\subsection{News for version 0.10.0}{
\subsection{codedoc::pkg_doc_fun}{

\code{codedoc::pkg_doc_fun} now a wrapper for \code{codedoc::pkg_doc_obj}.
}

\subsection{codedoc::pkg_doc_obj}{

New function \code{codedoc::pkg_doc_obj}.
}

}

\subsection{News for version 0.9.0}{
\subsection{codedoc::pkg_doc_fun}{

Fix for \code{codedoc::pkg_doc_fun} not being exported.
}

\subsection{codedoc::pkg_doc_package_description}{

Fix for \code{codedoc::pkg_doc_package_description} not being exported.
}

}

\subsection{News for version 0.8.0}{
\subsection{codedoc::extract_keyed_comment_blocks}{

\code{codedoc::extract_keyed_comment_blocks} default for \code{text_file_paths}
improved: It now excludes files ignored by \code{git} if it has been installed on
the system.
}

}

\subsection{News for version 0.7.0}{
\subsection{codedoc::extract_keyed_comment_blocks}{

\code{codedoc::extract_keyed_comment_blocks} default for \code{text_file_paths}
improved: It now excludes files under \verb{.../renv/library/} to avoid looking
through files in installed R packages. That is the typical directory for
packages installed by package \code{renv}.
}

}

\subsection{News for version 0.6.3}{
\subsection{codedoc::extract_keyed_comment_blocks}{

Fixed assertion on arg \code{detect_comment_lines}. Now allows for length > 1
\code{character} vectors as documented.
}

\subsection{codedoc::extract_keyed_comment_blocks}{

Fixed assertion on arg \code{clean_comment_lines}. Now allows for length > 1
\code{character} vectors as documented.
}

}

\subsection{News for version 0.6.2}{
\subsection{codedoc::pkg_doc_fun}{

Fixed \code{codedoc::pkg_doc_fun} extraction of param docs.
Multi-block docs could be mixed between parameters if they appeared in
a mixed order in text. Now they do not.
}

}

\subsection{News for version 0.6.1}{
\subsection{codedoc::pkg_doc_fun}{

\code{codedoc::pkg_doc_fun} now more robust when calling
\code{codedoc::extract_keyed_comment_blocks}.
}

}

\subsection{News for version 0.6.0}{
\subsection{codedoc::extract_keyed_comment_blocks}{

\code{codedoc::extract_keyed_comment_blocks} default for \code{text_file_paths}
expanded: Now uses regex
\code{"[.]((r)|(rmd)|(py)|(sql)|(cpp)|(hpp)|(c)|(h))$"} in \code{dir} call.
}

\subsection{codedoc::extract_keyed_comment_blocks}{

\code{codedoc::extract_keyed_comment_blocks} now replaces
\code{detect_comment_lines = NULL} with
\code{c(r = "^\\\\s*[#]+\\\\s?", rmd = "^\\\\s*[#]+\\\\s?", py = "^\\\\s*[#]+\\\\s?", sql = "^\\\\s*[-]{2,}\\\\s?", cpp = "^\\\\s*[/]{2,}\\\\s?", hpp = "^\\\\s*[/]{2,}\\\\s?", c = "^\\\\s*[/]{2,}\\\\s?", h = "^\\\\s*[/]{2,}\\\\s?")}.
}

\subsection{codedoc::extract_keyed_comment_blocks}{

\code{codedoc::extract_keyed_comment_blocks} default for \code{clean_comment_lines}
is (usually) what \code{detect_comment_lines} is.
}

\subsection{codedoc::pkg_doc_fun}{

New fun \code{codedoc::pkg_doc_fun}.
}

\subsection{codedoc::pkg_doc_package_description}{

New fun \code{codedoc::pkg_doc_package_description}.
}

}

\subsection{News for version 0.5.0}{
\subsection{codedoc}{

Replaced \verb{dbc::assert_is_data.frame*} -> \verb{dbc::assert_is_data_frame*}.
No effect on user, avoids a deprecated feature in \code{dbc}.
}

}

\subsection{News for version 0.4.1}{
\subsection{codedoc::codedoc_R_package_description}{

\verb{[codedoc::codedoc_R_package_description]} now uses field \code{Description}
from file \code{DESCRIPTION} if it exists as the, well, description if
\code{codedoc::codedoc_lines} failed to extract any lines.
}

}

\subsection{News for version 0.4.0}{
\subsection{codedoc::codedoc_news_df_to_markdown_news}{

\code{codedoc::codedoc_news_df_to_markdown_news} title format changed from
\verb{### Object x news:} to simply \verb{### x}.
}

}

\subsection{News for version 0.3.8}{
\subsection{codedoc::extract_keyed_comment_blocks}{

\verb{[codedoc:extract_keyed_comment_blocks]} default args changed:
\code{detect_comment_lines = "^\\\\s*[#*]\\\\s*"}
-> \code{detect_comment_lines = "^\\\\s*[#*]\\\\s?"}
\code{clean_comment_lines  = "^\\\\s*[#*]\\\\s?"}
-> \code{clean_comment_lines  = "^\\\\s*[#*]\\\\s*"}

This will improves retaining of indentation in comment blocks.
}

}

\subsection{News for version 0.3.7}{
\subsection{codedoc::extract_keyed_comment_blocks}{

\verb{[codedoc:extract_keyed_comment_blocks]} string interpolation improved.
If evaluating an expression results in an error, a more informative
error message is produced than previously. New error message includes
errored expression, original error message,
path to file producing the error and first line number of guilty
part of text.
}

}

\subsection{News for version 0.3.6}{
\subsection{codedoc::extract_keyed_comment_blocks}{

\verb{[codedoc:extract_keyed_comment_blocks]} arg \code{text_file_paths} default
improved --- now uses \code{pattern = "[.]((r)|(rmd))$"} in place of
\code{pattern = "((r)|(rmd))$"}.
}

\subsection{codedoc::extract_keyed_comment_blocks_}{

\verb{[codedoc:extract_keyed_comment_blocks_]} has been permanently removed.
Use \verb{[codedoc:extract_keyed_comment_blocks]}.
}

}

\subsection{News for version 0.3.5}{
\subsection{codedoc::extract_keyed_comment_blocks}{

\code{sort_by} arg now actually works.
}

}

\subsection{News for version 0.3.4}{
\subsection{codedoc::extract_keyed_comment_blocks}{

User can now easily sort output \code{block_df} by columns of their choice via
new arg \code{sort_by}.
}

}

\subsection{News for version 0.3.3}{
\subsection{codedoc::extract_keyed_comment_blocks}{

The \code{data.frame} of collected blocks is now always sorted by
\code{text_file_path} and \code{first_block_line}. The order of results was this
also before, but now it has been made explicit in code and documentation.
}

\subsection{codedoc::extract_keyed_comment_blocks_}{

\verb{[codedoc:extract_keyed_comment_blocks_]} no longer usable and always
raises an error. Use \verb{[codedoc:extract_keyed_comment_blocks]}.
}

}

\subsection{News for version 0.3.2}{
\subsection{codedoc::extract_keyed_comment_blocks}{

\code{codedoc::extract_keyed_comment_block} ensured to return a proper empty
data.frame when nothing was extracted. Column \code{comment_block} was
previously missing.
}

}

\subsection{News for version 0.3.1}{
\subsection{codedoc::codedoc_R_package_description}{

\verb{[codedoc::codedoc_R_package_description]} now includes a markdown
comment in the \code{.md} which indicates that the file was generated
by \code{codedoc}.
}

\subsection{codedoc::codedoc_news_df_to_markdown_news}{

\verb{[codedoc::codedoc_news_df_to_markdown_news]} now includes a markdown
comment to notify that its output was generated by \code{codedoc}.
}

\subsection{codedoc::codedoc_news_df_to_roxygen_news_by_version}{

\verb{[codedoc::codedoc_news_df_to_roxygen_news_by_version]}:
The order of versions in is now descending instead of ascending.
}

\subsection{codedoc::codedoc_news_for_R_package}{

\verb{[codedoc::codedoc_news_for_R_package]} avoids markdown comments in
output (but they are included in \code{NEWS.md}).
}

}

\subsection{News for version 0.3.0}{
\subsection{codedoc::extract_keyed_comment_blocks}{

\verb{[codedoc:extract_keyed_comment_blocks]} gained arg \code{assertion_type}.
}

\subsection{codedoc::extract_keyed_comment_blocks_}{

\verb{[codedoc:extract_keyed_comment_blocks_]} marked for deprecation.
Use \verb{[codedoc:extract_keyed_comment_blocks]}.
}

\subsection{codedoc::render_codedoc}{

\verb{[codedoc:render_codedoc]} gained arg \code{assertion_type}.
}

\subsection{codedoc::render_codedoc_}{

\verb{[codedoc:render_codedoc_]} marked for deprecation.
Use \verb{[codedoc:render_codedoc]}.
}

}

\subsection{News for version 0.2.15}{
\subsection{codedoc::codedoc_lines}{

New exported function \verb{[codedoc::codedoc_lines]}.
}

\subsection{codedoc::codedoc_markdown_news}{

New exported function \code{codedoc::codedoc_markdown_news}.
}

\subsection{codedoc::codedoc_news_df}{

New exported function \verb{[codedoc::codedoc_news_df]}.
}

\subsection{codedoc::codedoc_news_df_to_markdown_news}{

New exported function \code{codedoc::codedoc_news_df_to_markdown_news}.
}

\subsection{codedoc::codedoc_news_df_to_roxygen_news_by_object_name}{

New exported function \code{codedoc::codedoc_news_df_to_roxygen_news_by_object_name}.
}

\subsection{codedoc::codedoc_news_df_to_roxygen_news_by_object_name}{

New exported function \code{codedoc::codedoc_roxygen_news_by_object_name}.
}

\subsection{codedoc::codedoc_news_df_to_roxygen_news_by_version}{

New exported function \code{codedoc::codedoc_news_df_to_roxygen_news_by_version}.
}

\subsection{codedoc::codedoc_roxygen_news_by_version}{

New exported function \code{codedoc::codedoc_roxygen_news_by_version}.
}

\subsection{codedoc::extract_keyed_comment_blocks}{

Changes to function \verb{[codedoc::extract_keyed_comment_blocks]} arguments:
\itemize{
\item \code{text_file_paths}: allowed to also be \code{NULL}; default now \code{NULL}
\item \code{detect_comment_lines}: allowed to also be \code{character}; default now
\code{"^\\\\s*[#*]\\\\s*"}
\item \code{clean_comment_lines}: allowed to also be \code{character}; default now
\code{"^\\\\s*[#*]\\\\s*"}
\item \code{detect_allowed_keys}: allowed to also be \code{character}; default now
\code{""}
}
}

\subsection{codedoc::extract_keyed_comment_blocks_}{

Changes to function \verb{[codedoc::extract_keyed_comment_blocks]} arguments:
\itemize{
\item \code{text_file_paths}: allowed to also be \code{NULL}; default now \code{NULL}
\item \code{detect_comment_lines}: allowed to also be \code{character}; default now
\code{"^\\\\s*[#*]\\\\s*"}
\item \code{clean_comment_lines}: allowed to also be \code{character}; default now
\code{"^\\\\s*[#*]\\\\s*"}
\item \code{detect_allowed_keys}: allowed to also be \code{character}; default now
\code{""}
}
}

}
}

\keyword{internal}
