% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codedoc_package.R
\docType{package}
\name{codedoc}
\alias{codedoc}
\title{codedoc: Functions Document Your Code}
\description{
\code{codedoc} is designed to aid your code documentation. Extract comments
from the midst of your code and process them into documentation.
}
\section{News}{

\subsection{News for version 0.2.15}{
\subsection{Object codedoc::codedoc_lines news:}{

New exported function \verb{[codedoc::codedoc_lines]}.
}

\subsection{Object codedoc::codedoc_markdown_news news:}{

New exported function \code{codedoc::codedoc_markdown_news}.
}

\subsection{Object codedoc::codedoc_news_df news:}{

New exported function \verb{[codedoc::codedoc_news_df]}.
}

\subsection{Object codedoc::codedoc_news_df_to_markdown_news news:}{

New exported function \code{codedoc::codedoc_news_df_to_markdown_news}.
}

\subsection{Object codedoc::codedoc_news_df_to_roxygen_news_by_object_name news:}{

New exported function \code{codedoc::codedoc_news_df_to_roxygen_news_by_object_name}.
}

\subsection{Object codedoc::codedoc_news_df_to_roxygen_news_by_object_name news:}{

New exported function \code{codedoc::codedoc_roxygen_news_by_object_name}.
}

\subsection{Object codedoc::codedoc_news_df_to_roxygen_news_by_version news:}{

New exported function \code{codedoc::codedoc_news_df_to_roxygen_news_by_version}.
}

\subsection{Object codedoc::codedoc_roxygen_news_by_version news:}{

New exported function \code{codedoc::codedoc_roxygen_news_by_version}.
}

\subsection{Object codedoc::extract_keyed_comment_blocks news:}{

Changes to function \verb{[codedoc::extract_keyed_comment_blocks]} arguments:
\itemize{
\item \code{text_file_paths}: allowed to also be \code{NULL}; default now \code{NULL}
\item \code{detect_comment_lines}: allowed to also be \code{character}; default now
\code{"^\\\\s*[#*]\\\\s*"}
\item \code{clean_comment_lines}: allowed to also be \code{character}; default now
\code{"^\\\\s*[#*]\\\\s*"}
\item \code{detect_allowed_keys}: allowed to also be \code{character}; default now
\code{""}
}
}

\subsection{Object codedoc::extract_keyed_comment_blocks_ news:}{

Changes to function \verb{[codedoc::extract_keyed_comment_blocks]} arguments:
\itemize{
\item \code{text_file_paths}: allowed to also be \code{NULL}; default now \code{NULL}
\item \code{detect_comment_lines}: allowed to also be \code{character}; default now
\code{"^\\\\s*[#*]\\\\s*"}
\item \code{clean_comment_lines}: allowed to also be \code{character}; default now
\code{"^\\\\s*[#*]\\\\s*"}
\item \code{detect_allowed_keys}: allowed to also be \code{character}; default now
\code{""}
}
}

}
}

