% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codedoc_lines.R
\name{codedoc_lines}
\alias{codedoc_lines}
\title{\code{codedoc} Comment Block Lines}
\usage{
codedoc_lines(
  detect_allowed_keys,
  text_file_paths = NULL,
  extract_arg_list = NULL,
  assertion_type = "input"
)
}
\arguments{
\item{detect_allowed_keys}{see \verb{[codedoc::extract_keyed_comment_blocks]}}

\item{text_file_paths}{see \verb{[codedoc::extract_keyed_comment_blocks]}}

\item{extract_arg_list}{\itemize{
\item \code{NULL}: no additional args
\item \code{list}: these args are passed to \verb{[codedoc::extract_keyed_comment_blocks]}
}}

\item{assertion_type}{see \verb{[dbc::report_to_assertion]}}
}
\description{
Tools to extract comment blocks and return a vector of strings (lines).
}
\section{Functions}{

}

