% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codedoc_r_package_description.R
\name{codedoc_R_package_description}
\alias{codedoc_R_package_description}
\title{\code{codedoc} to R Package Description}
\usage{
codedoc_R_package_description(
  R_package_name,
  text_file_paths = NULL,
  extract_arg_list = NULL,
  assertion_type = "input"
)
}
\arguments{
\item{R_package_name}{\verb{[character]} (no default)

Turned into \code{detect_allowed_keys} arg for \verb{[codedoc::codedoc_lines]} via

\if{html}{\out{<div class="sourceCode r">}}\preformatted{detect_allowed_keys <- paste0(
  "^\\\\QR_package_description(", R_package_name, ")\\\\E$"
)
}\if{html}{\out{</div>}}}

\item{text_file_paths}{\verb{[character]} see \verb{[codedoc::codedoc_lines]}}

\item{extract_arg_list}{\verb{[character]} see \verb{[codedoc::codedoc_lines]}}

\item{assertion_type}{see \verb{[dbc::report_to_assertion]}}
}
\description{
Extract \code{codedoc} comment block(s) and turn them into R package
description help page (roxygen) and a \code{README.md}.
}
\section{News for version 0.3.1}{


\verb{[codedoc::codedoc_R_package_description]} now includes a markdown
comment in the \code{.md} which indicates that the file was generated
by \code{codedoc}.
}

