% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{example_template}
\alias{example_template}
\alias{example_template_dir_path}
\alias{example_template_file_names}
\alias{example_template_file_paths}
\alias{example_template_file_path}
\alias{example_template_lines}
\title{Example Templates}
\usage{
example_template_dir_path()

example_template_file_names()

example_template_file_paths()

example_template_file_path(example_template_file_name)

example_template_lines(example_template_file_name)
}
\arguments{
\item{example_template_file_name}{\verb{[character]} (mandatory, no default)

name of example template file}
}
\description{
Retrieve example codedoc templates.
}
\details{
\itemize{
\item \code{example_template_dir_path} gets the path to the dir containing example
templates in this package
}

\itemize{
\item \code{example_template_file_names} gets the names example template files
}

\itemize{
\item \code{example_template_file_paths} gets the full paths to example template files
}

\itemize{
\item \code{example_template_file_path} gets the path to an example template based on
its name
}

\itemize{
\item \code{example_template_lines} reads the lines of an example template based on
its name
}
}
