% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\name{example_text_file}
\alias{example_text_file}
\alias{example_text_file_dir}
\alias{example_text_file_names}
\alias{example_text_file_paths}
\alias{example_text_file_path}
\alias{example_text_file_lines}
\title{Example Text Files}
\usage{
example_text_file_dir()

example_text_file_names()

example_text_file_paths()

example_text_file_path(example_text_file_name)

example_text_file_lines(example_text_file_name)
}
\arguments{
\item{example_text_file_name}{\verb{[character]} (mandatory, no default)

name of example text file}
}
\description{
Retrieve example text files processable by codedoc.
}
\details{
\itemize{
\item \code{example_text_file_dir} gets the path to the dir containing example text
files
}

\itemize{
\item \code{example_text_file_names} gets the names of all example text files in
this package
}

\itemize{
\item \code{example_text_file_paths} gets the full paths to all example text files in
this package
}

\itemize{
\item \code{example_text_file_path} gets the full path to an example text file
based on its name
}

\itemize{
\item \code{example_text_file_lines} reads the contents of an example text file
into R based on its name
}
}
