% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_doc.R
\name{pkg_doc}
\alias{pkg_doc}
\alias{pkg_doc_obj}
\alias{pkg_doc_fun}
\alias{pkg_doc_package_description}
\title{Document R Package}
\usage{
pkg_doc_obj(regex, rdname = NULL, text_file_paths = NULL)

pkg_doc_fun(regex, rdname = NULL, text_file_paths = NULL)

pkg_doc_package_description(
  text_file_paths = NULL,
  desc_arg_list = NULL,
  news_arg_list = NULL
)
}
\arguments{
\item{regex}{\verb{[character]} (no default)

Any regular expression accepted by \code{grepl} that contains \code{"::"} or \code{":::"}.
E.g. \code{"mypkg::myfun"}.}

\item{rdname}{\verb{[NULL, character]} (default \code{NULL})
\itemize{
\item \code{NULL}: No effect on generated docs.
\item \code{character}: Generate additional line \code{paste0("@rdname ", rdname)}.
}}

\item{text_file_paths}{\verb{[NULL, character]} (default \code{NULL})
\itemize{
\item \code{codedoc::pkg_doc_package_description}:
Overrides \code{text_file_paths} passed to
\code{codedoc::codedoc_R_package_description} and
\code{codedoc::codedoc_news_for_R_package}.
}}

\item{desc_arg_list}{\verb{[NULL, list]} (default \code{NULL})

List of arguments passed to
\code{codedoc::codedoc_R_package_description}.}

\item{news_arg_list}{\verb{[NULL, list]} (default \code{NULL})

List of arguments passed to
\code{codedoc::codedoc_news_for_R_package}.}
}
\value{
\strong{codedoc::pkg_doc_obj}

Returns a \code{character} vector of lines for further processing by \code{roxygen2}.

\strong{codedoc::pkg_doc_package_description}

Returns a \code{character} vector, the combined outputs of the functions called
by this one.
}
\description{
Tools to document an R package.
}
\section{Functions}{

\strong{codedoc::pkg_doc_obj}

Document an object in an R package. Performs the following steps:
\itemize{
\item Calls \verb{[codedoc::extract_keyed_comment_blocks]}.
\item Collect comment block lines whose keys match one of:
\itemize{
\item \code{"^\%s::"} for arguments --- but you should use roxygen
block directly unless you have a specific reason.
\item \code{"^\%s$"} for the description of the function.
These go into either \code{Details} or section \code{Functions}, depending on
\code{rdname}.
\item \code{"^return[(]\%s[)]$"} for docs describing what the
function returns. These appear under the section \code{Value} in the help
page.
\item \code{"^examples[(]\%s[)]$"} for the Examples section.
}
\item Call \verb{[codedoc::codedoc_roxygen_news_by_version]} to extract news for
the function, if any.
\item If \code{regex} looks something like \code{xx::xx_make_column_xxxx}, matching
regular expression
\verb{^[a-zA-Z0-9._-]+::[a-zA-Z0-9._-]+_make_column_[a-zA-Z0-9._-]+$},
its argument names which do not contain the words \code{dataset} or \code{setting}
are assumed to be column names and are documented simply with
c("@param some.column (no default)", "", "Column some_column values.").
This means you don't have to document such column arguments separately.
\item Finally, collect all docs into one long string vector and add \verb{@export}
to the beginning of the vector.
Document a function in an R package. Wrapper for
\code{codedoc::pkg_doc_fun} .
}

[)]: R:)
\link{(}: R:(
[)]: R:)
[codedoc::codedoc_roxygen_news_by_version]: R:codedoc::codedoc_roxygen_news_by_version
[a-zA-Z0-9.\emph{-]: R:a-zA-Z0-9.}-
[a-zA-Z0-9.\emph{-]: R:a-zA-Z0-9.}-
[a-zA-Z0-9.\emph{-]: R:a-zA-Z0-9.}-


\strong{codedoc::pkg_doc_package_description}

Calls \code{codedoc::codedoc_R_package_description} and
\code{codedoc::codedoc_news_for_R_package}.
}

\section{News for version 0.10.4}{


Add handling for examples. E.g. key \code{examples(mypkg::myfun)}.

\code{codedoc::pkg_doc_obj} argument \code{grepl_arg_list} removed because it was
not worth the trouble. To have a fixed expression, instead of passing
\code{fixed = TRUE} one can do \code{"\\\\Qmypkg::myobj\\\\E"}.



\code{codedoc::pkg_doc_fun} arg \code{grepl_arg_list} removed.
}

\section{News for version 0.10.3}{


\code{codedoc::pkg_doc_obj} argument \code{grepl_arg_list} handling improved:
\itemize{
\item By default \code{perl = TRUE} but user can override this.
\item If user supplies \code{fixed = TRUE}, we always set \code{perl = FALSE}.
}
}

\section{News for version 0.10.2}{


\code{codedoc::pkg_doc_obj} gains new arg \code{grepl_arg_list}.



\code{codedoc::pkg_doc_fun} gains new arg \code{grepl_arg_list}.
}

\section{News for version 0.10.1}{


Expand \code{codedoc::pkg_doc_obj} allowed \code{regex} patterns.
Now \code{regex} can be any string that \code{grepl} can handle.
}

\section{News for version 0.10.0}{


New function \code{codedoc::pkg_doc_obj}.



\code{codedoc::pkg_doc_fun} now a wrapper for \code{codedoc::pkg_doc_obj}.
}

\section{News for version 0.9.0}{


Fix for \code{codedoc::pkg_doc_fun} not being exported.



Fix for \code{codedoc::pkg_doc_package_description} not being exported.
}

\section{News for version 0.6.2}{


Fixed \code{codedoc::pkg_doc_fun} extraction of param docs.
Multi-block docs could be mixed between parameters if they appeared in
a mixed order in text. Now they do not.
}

\section{News for version 0.6.1}{


\code{codedoc::pkg_doc_fun} now more robust when calling
\code{codedoc::extract_keyed_comment_blocks}.
}

\section{News for version 0.6.0}{


New fun \code{codedoc::pkg_doc_fun}.



New fun \code{codedoc::pkg_doc_package_description}.
}

