% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_raw.R
\name{read_raw_codedocs}
\alias{read_raw_codedocs}
\title{Read codedoc-formatted Code Comments}
\usage{
read_raw_codedocs(
  text_file_path,
  detect_comment_lines = function(x) grepl("^\\\\s*[#*]\\\\s*", x),
  clean_comment_lines = function(x) sub("^\\\\s*[#*]\\\\s*", "", x),
  readLines_arg_list = list()
)
}
\arguments{
\item{text_file_path}{\verb{[character]} (mandatory, no default)

path to a text file; the file is not inspected in any way to be a text file
based on file extension or any other characteristics, it is merely assumed
to be one}

\item{detect_comment_lines}{\verb{[function]}
(optional, default \code{function(x) grepl("^\\\\s*[#*]", x)})

a function which takes lines of text as input and outputs a logical vector
of the same length as input which is \code{TRUE} when the line is a comment line;
the default detects lines that start with the regex \code{"^\\\\s*[#*]\\\\s*"}}

\item{clean_comment_lines}{\verb{[function]}
(optional, default \code{function(x) sub("^\\\\s*[#*]\\\\s*", "", x)})

a function which takes lines of text as input and outputs a character vector
of the same length as input; this function takes comment lines and should
strip any comment characters preceding the text itself;
the default removes any substrings matching regex \code{"^\\\\s*[#*]\\\\s*"}}

\item{readLines_arg_list}{\verb{[list]} (mandatory, default \code{list()})

list of arguments passed to \link{readLines}; \code{con} is always set to
\code{text_file_path}}
}
\description{
Extracts blocks of specifically formatted comments from text file.
}
