% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_codedoc.R
\docType{package}
\name{codedoc}
\alias{codedoc}
\title{codedoc: Functions Document Your Code}
\description{
\code{codedoc} is designed to aid your code documentation. Extract comments
from the midst of your code and process them into documentation.
}
\section{News}{

\subsection{News for version 0.3.3}{
\subsection{Object codedoc::extract_keyed_comment_blocks news:}{

The \code{data.frame} of collected blocks is now always sorted by
\code{text_file_path} and \code{first_block_line}. The order of results was this
also before, but now it has been made explicit in code and documentation.
}

\subsection{Object codedoc::extract_keyed_comment_blocks_ news:}{

\verb{[codedoc:extract_keyed_comment_blocks_]} no longer usable and always
raises an error. Use \verb{[codedoc:extract_keyed_comment_blocks]}.
}

}

\subsection{News for version 0.3.2}{
\subsection{Object codedoc::extract_keyed_comment_blocks news:}{

\code{codedoc::extract_keyed_comment_block} ensured to return a proper empty
data.frame when nothing was extracted. Column \code{comment_block} was
previously missing.
}

}

\subsection{News for version 0.3.1}{
\subsection{Object codedoc::codedoc_R_package_description news:}{

\verb{[codedoc::codedoc_R_package_description]} now includes a markdown
comment in the \code{.md} which indicates that the file was generated
by \code{codedoc}.
}

\subsection{Object codedoc::codedoc_news_df_to_markdown_news news:}{

\verb{[codedoc::codedoc_news_df_to_markdown_news]} now includes a markdown
comment to notify that its output was generated by \code{codedoc}.
}

\subsection{Object codedoc::codedoc_news_df_to_roxygen_news_by_version news:}{

\verb{[codedoc::codedoc_news_df_to_roxygen_news_by_version]}:
The order of versions in is now descending instead of ascending.
}

\subsection{Object codedoc::codedoc_news_for_R_package news:}{

\verb{[codedoc::codedoc_news_for_R_package]} avoids markdown comments in
output (but they are included in \code{NEWS.md}).
}

}

\subsection{News for version 0.3.0}{
\subsection{Object codedoc::extract_keyed_comment_blocks news:}{

\verb{[codedoc:extract_keyed_comment_blocks]} gained arg \code{assertion_type}.
}

\subsection{Object codedoc::extract_keyed_comment_blocks_ news:}{

\verb{[codedoc:extract_keyed_comment_blocks_]} marked for deprecation.
Use \verb{[codedoc:extract_keyed_comment_blocks]}.
}

\subsection{Object codedoc::render_codedoc news:}{

\verb{[codedoc:render_codedoc]} gained arg \code{assertion_type}.
}

\subsection{Object codedoc::render_codedoc_ news:}{

\verb{[codedoc:render_codedoc_]} marked for deprecation.
Use \verb{[codedoc:render_codedoc]}.
}

}

\subsection{News for version 0.2.15}{
\subsection{Object codedoc::codedoc_lines news:}{

New exported function \verb{[codedoc::codedoc_lines]}.
}

\subsection{Object codedoc::codedoc_markdown_news news:}{

New exported function \code{codedoc::codedoc_markdown_news}.
}

\subsection{Object codedoc::codedoc_news_df news:}{

New exported function \verb{[codedoc::codedoc_news_df]}.
}

\subsection{Object codedoc::codedoc_news_df_to_markdown_news news:}{

New exported function \code{codedoc::codedoc_news_df_to_markdown_news}.
}

\subsection{Object codedoc::codedoc_news_df_to_roxygen_news_by_object_name news:}{

New exported function \code{codedoc::codedoc_news_df_to_roxygen_news_by_object_name}.
}

\subsection{Object codedoc::codedoc_news_df_to_roxygen_news_by_object_name news:}{

New exported function \code{codedoc::codedoc_roxygen_news_by_object_name}.
}

\subsection{Object codedoc::codedoc_news_df_to_roxygen_news_by_version news:}{

New exported function \code{codedoc::codedoc_news_df_to_roxygen_news_by_version}.
}

\subsection{Object codedoc::codedoc_roxygen_news_by_version news:}{

New exported function \code{codedoc::codedoc_roxygen_news_by_version}.
}

\subsection{Object codedoc::extract_keyed_comment_blocks news:}{

Changes to function \verb{[codedoc::extract_keyed_comment_blocks]} arguments:
\itemize{
\item \code{text_file_paths}: allowed to also be \code{NULL}; default now \code{NULL}
\item \code{detect_comment_lines}: allowed to also be \code{character}; default now
\code{"^\\\\s*[#*]\\\\s*"}
\item \code{clean_comment_lines}: allowed to also be \code{character}; default now
\code{"^\\\\s*[#*]\\\\s*"}
\item \code{detect_allowed_keys}: allowed to also be \code{character}; default now
\code{""}
}
}

\subsection{Object codedoc::extract_keyed_comment_blocks_ news:}{

Changes to function \verb{[codedoc::extract_keyed_comment_blocks]} arguments:
\itemize{
\item \code{text_file_paths}: allowed to also be \code{NULL}; default now \code{NULL}
\item \code{detect_comment_lines}: allowed to also be \code{character}; default now
\code{"^\\\\s*[#*]\\\\s*"}
\item \code{clean_comment_lines}: allowed to also be \code{character}; default now
\code{"^\\\\s*[#*]\\\\s*"}
\item \code{detect_allowed_keys}: allowed to also be \code{character}; default now
\code{""}
}
}

}
}

