% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_doc.R
\name{pkg_doc}
\alias{pkg_doc}
\alias{pkg_doc_fun}
\alias{pkg_doc_package_description}
\title{Document R Package}
\usage{
pkg_doc_fun(regex, rdname = NULL, text_file_paths = NULL)

pkg_doc_package_description(
  text_file_paths = NULL,
  desc_arg_list = NULL,
  news_arg_list = NULL
)
}
\arguments{
\item{regex}{\verb{[character]} (no default)

Must look like e.g. \code{"mypkg::myfun"}, where you write
\code{codedoc} comment blocks with \code{mypkg::myfun} as/in the keys.
This is used to detect which \code{codedoc} comment blocks to use in
generating documents. Passed to \code{codedoc::codedoc_lines} as
both \code{paste0("^", regex, "::")} and \code{paste0("^", regex, "$")} and to
\code{codedoc::codedoc_roxygen_news_by_version} as-is.
\code{paste0("^", regex, "::")} is meant for detecting \code{codedoc} comment
blocks for arguments, e.g. \verb{mypkgs::myfun::arg} --- but you should simply
use \code{roxygen} blocks normally unless there is a special reason to do
otherwise.}

\item{rdname}{\verb{[NULL, character]} (default \code{NULL})
\itemize{
\item \code{NULL}: No effect on generated docs.
\item \code{character}: Generate additional line \code{paste0("@rdname ", rdname)}.
}}

\item{text_file_paths}{\verb{[NULL, character]} (default \code{NULL})
\itemize{
\item \code{codedoc::pkg_doc_package_description}:
Overrides \code{text_file_paths} passed to
\code{codedoc::codedoc_R_package_description} and
\code{codedoc::codedoc_news_for_R_package}.
}}

\item{desc_arg_list}{\verb{[NULL, list]} (default \code{NULL})

List of arguments passed to
\code{codedoc::codedoc_R_package_description}.}

\item{news_arg_list}{\verb{[NULL, list]} (default \code{NULL})

List of arguments passed to
\code{codedoc::codedoc_news_for_R_package}.}
}
\value{
\strong{codedoc::pkg_doc_fun}

Returns a \code{character} vector of lines for further processing by \code{roxygen2}.

\strong{codedoc::pkg_doc_package_description}

Returns a \code{character} vector, the combined outputs of the functions called
by this one.
}
\description{
Tools to document an R package.
}
\section{Functions}{

\strong{codedoc::pkg_doc_fun}

Document a function in an R package.

Performs the following steps.
\itemize{
\item Calls \verb{[codedoc::extract_keyed_comment_blocks]}.
\item Collect comment block lines whose keys match one of:
}
\itemize{
\item \code{paste0("^", regex, "::")} for arguments --- but you should use roxygen
block directly unless you have a specific reason.
\item \code{paste0("^", regex, "$")} for the description of the function.
These go into either \code{Details} or section \code{Functions}, depending on
\code{rdname}.
\item \code{paste0("^return[(]", regex, "[)]$")} for docs describing the what the
function returns. These appear under the section \code{Value} in the help
page.
}
\itemize{
\item Call \verb{[codedoc::codedoc_roxygen_news_by_version]} to extract news for
the function, if any.
\item If \code{regex} looks something like \code{xx::xx_make_column_xxxx}, matching
regular expression
\verb{^[a-zA-Z0-9._-]+::[a-zA-Z0-9._-]+_make_column_[a-zA-Z0-9._-]+$},
its argument names which do not contain the words \code{dataset} or \code{setting}
are assumed to be column names and are documented simply with
c("@param some.column (no default)", "", "Column some_column values.").
This means you don't have to document such column arguments separately.
\item Finally, collect all docs into one long string vector and add \verb{@export}
to the beginning of the vector.
}

[)]: R:)
[codedoc::codedoc_roxygen_news_by_version]: R:codedoc::codedoc_roxygen_news_by_version
[a-zA-Z0-9.\emph{-]: R:a-zA-Z0-9.}-
[a-zA-Z0-9.\emph{-]: R:a-zA-Z0-9.}-
[a-zA-Z0-9.\emph{-]: R:a-zA-Z0-9.}-


\strong{codedoc::pkg_doc_package_description}

Calls \code{codedoc::codedoc_R_package_description} and
\code{codedoc::codedoc_news_for_R_package}.
}

\section{News for version 0.6.0}{


New fun \code{codedoc::pkg_doc_fun}.



New fun \code{codedoc::pkg_doc_package_description}.
}

