% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{render_codedoc}
\alias{render_codedoc}
\title{Render Code Documentation}
\usage{
render_codedoc(
  key_df,
  template_file_path = NULL,
  writeLines_arg_list = list(),
  render_arg_list = list()
)
}
\arguments{
\item{key_df}{\verb{[data.frame]} (mandatory, no default)

a data.frame which must have columns \code{key}, \code{comment_block}, and
\code{text_file_path} as in the output of \link{extract_keyed_comment_blocks}}

\item{template_file_path}{\verb{[NULL, character]} (optional, default \code{NULL})
\itemize{
\item \code{NULL}: no template is used; see section Template
\item \code{character}: path to a template file; see section Template
}}

\item{writeLines_arg_list}{\verb{[list]} (optional, default \code{list()})

list of arguments passed to \code{writeLines} when writing temporary .rmd
file to render; \code{con} and \code{text} are always determined internally}

\item{render_arg_list}{\verb{[list]} (optional, default \code{list()})

list of arguments passed to \link[rmarkdown:render]{rmarkdown::render}; \code{input} is always
determined internally; the default is\preformatted{list(output_file = "codedoc.md",
     output_dir = getwd(),
     quiet = TRUE)
}}
}
\description{
Render using \link[rmarkdown:render]{rmarkdown::render} a documentation file using extracted
code comment blocks.
}
\section{Template}{


TODO
}

