% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{render_codedoc}
\alias{render_codedoc}
\title{Render Code Documentation}
\usage{
render_codedoc(
  key_df,
  template_file_path = NULL,
  writeLines_arg_list = list(),
  render_arg_list = list()
)
}
\arguments{
\item{key_df}{\verb{[data.frame]} (mandatory, no default)

a data.frame which must have columns \code{key}, \code{comment_block}, and
\code{text_file_path} as in the output of \link{extract_keyed_comment_blocks}}

\item{template_file_path}{\verb{[NULL, character]} (optional, default \code{NULL})
\itemize{
\item \code{NULL}: no template is used; see section Template
\item \code{character}: path to a template file; see section Template
}}

\item{writeLines_arg_list}{\verb{[list]} (optional, default \code{list()})

list of arguments passed to \code{writeLines} when writing temporary .rmd
file to render; \code{con} and \code{text} are always determined internally}

\item{render_arg_list}{\verb{[list]} (optional, default \code{list()})

list of arguments passed to \link[rmarkdown:render]{rmarkdown::render}; \code{input} is always
determined internally; the default settings set internally in this function
are\preformatted{list(output_file = "codedoc.md",
     output_format = "md_document",
     output_dir = getwd(),
     quiet = TRUE,
     envir = new.env(parent = parent.frame(1L)))
}

where \code{envir} will by default be the environment where \code{render_codedoc} was
called. \code{envir} is additionally always populated by these objects:
\itemize{
\item \code{key_df}: the user-supplied arg
\item \code{codedoc_lines}: a function with only the argument \code{key}, which must be
one of the keys in \code{key_df}; the comment lines are returned for the
corresponding key as a character vector
\item \code{codedoc_text}: as \code{codedoc_lines}. but instead of the lines as a character
vector (with one or more elements) this function returns exactly one
string, formed by pasting the lines together with
\code{paste0(lines, collapse = "\\n")}; this is more convenient for rendering
text in the template (see section Template)
}}
}
\description{
Render using \link[rmarkdown:render]{rmarkdown::render} a documentation file using extracted
code comment blocks.
}
\section{Template}{


The template file must be an Rmarkdown file (see e.g. \link[rmarkdown:render]{rmarkdown::render}).
It's contents are completely up to you. You should make use the objects
created into the rendering environment (see \code{envir} in \link[rmarkdown:render]{rmarkdown::render})
by this function; see arg \code{render_arg_list}. The template should at a minimum
have the Rmarkdown header with a title, e.g.\preformatted{---
title: My title
---
}

See function \link{example_template} for examples.
}

